requirement.txt
fastapi==0.101.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
httpx==0.24.1
openai==1.31.0
pydantic==2.5.3
